% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTshinyModule.R
\name{DTmodule}
\alias{DTmodule}
\title{DTmodule}
\usage{
DTmodule(input, output, session, data, checkbox.colname = "Select",
  checked.rows.id.prefix = "Row")
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{data}{}

\item{checkbox.colname}{}

\item{checked.rows.id.prefix}{}
}
\value{
reactiveValues
}
\description{
DTmodule
}
