% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{runExample}
\usage{
runExample(example = "01_simpleApp")
}
\arguments{
\item{example}{}
}
\description{
runExample
}
\examples{
\dontrun{
DTshinyModule::runExample("01_simpleApp")
}
}
