% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTshinyModule.R
\name{DTmoduleUI}
\alias{DTmoduleUI}
\alias{DTmodule}
\title{DTmoduleUI}
\usage{
DTmoduleUI(id)

DTmodule(input, output, session, data, checkbox.colname = "Select",
  actions.colname = "Actions", checked.rows.id.prefix = "Row",
  header = NULL, add.row = TRUE, delete.row = TRUE,
  is.checked.rule = NULL)
}
\arguments{
\item{id}{}

\item{input}{}

\item{output}{}

\item{session}{}

\item{data}{}

\item{checkbox.colname}{}

\item{actions.colname}{}

\item{checked.rows.id.prefix}{}

\item{header}{}

\item{add.row}{}

\item{delete.row}{}

\item{is.checked.rule}{}

\item{...}{}
}
\value{
reactiveValues
}
\description{
DTmoduleUI

DTmodule
}
\author{
Kamil Foltyński

Kamil Foltyński
}
